% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{calculate_kpi_score}
\alias{calculate_kpi_score}
\title{Get KPI Score}
\usage{
calculate_kpi_score(
  data,
  sector_variable,
  sector_profiles,
  probs = c(0, 0.25, 0.5, 0.75, 1),
  score_table = c(0, 5, 20, 20, 5, 0),
  debug = FALSE
)
}
\arguments{
\item{data}{A list of data to calculate score, each item present a data frame of KPI values}

\item{sector_variable}{Sector variable name}

\item{sector_profiles}{Sector profile}

\item{probs}{List of quantiles, default is probs = c(0, 0.25, 0.50, 0.75, 1)}

\item{score_table}{List of score table corresponding with probs, default is score_table = c(0, 5, 20, 20, 5, 0)}

\item{debug}{Debug mode}
}
\value{
A list of data scores, each item present a data frame of KPI scores
}
\description{
Get KPI Score
}
\examples{
sector_variable <- "NGANHKT"
probs <- c(0, 0.05, 0.95, 1)
score_table <- c(0, 5, 50, 5)
discrete_static_kpi_names <- preprocessed$discrete_static_variables
data_scores <- tfdbn::calculate_kpi_score(test_data, sector_variable, sector_profiles, probs, score_table)

data_id <- names(data_scores)
for(id in data_id) {
  anomally_kpi <- get_anomally_kpi(data_scores[[id]], 5)
}
}
