% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{calculate_prob_score}
\alias{calculate_prob_score}
\title{Get anomaly KPI probabilities}
\usage{
calculate_prob_score(
  fitted,
  data_test,
  data_score,
  sector_variable,
  except_variables,
  sector_factor,
  sector_profile,
  kpi_score_threshold,
  probs,
  n_cluster = 4,
  n_times = 10,
  debug = FALSE
)
}
\arguments{
\item{fitted}{A list of data to calculate score, each item present a data frame of KPI values}

\item{data_test}{A list of data to calculate score, each item present a data frame of KPI values}

\item{data_score}{A list of data to calculate score, each item present a data frame of KPI scores}

\item{sector_variable}{Sector variable name}

\item{except_variables}{List of variables to ignore}

\item{sector_factor}{Sector in string}

\item{sector_profile}{Sector KPI profile}

\item{kpi_score_threshold}{Threshold}

\item{probs}{List of quantiles, default is probs = c(0, 0.25, 0.50, 0.75, 1)}

\item{n_cluster}{Number of CPU to calculate}

\item{n_times}{Number of calculation}

\item{debug}{Debug mode}
}
\value{
A list of data scores, each item present a data frame of KPI scores
}
\description{
Get anomaly KPI probabilities
}
\examples{
sector_variable <- "NGANHKT"

data_probs <- tfdbn::calculate_prob_score(fitted, test_data, data_scores, sector_variable, c("VON", "TUOI"), data_factor[[sector_variable]], sector_profiles, 5, c(0, 0.05, 0.95, 1), 4, 5)
}
