% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{preprocess_training_data}
\alias{preprocess_training_data}
\title{Pre-processing data before training with DBN}
\usage{
preprocess_training_data(
  data,
  type,
  index_column,
  time_column,
  time_values,
  continuous_kpi_names,
  continuous_static_kpi_names,
  discrete_static_kpi_names,
  current_layers,
  desire_layers,
  quantile_number = -1,
  na_omit = TRUE,
  debug = FALSE
)
}
\arguments{
\item{data}{A data frame which rows are financial reports and columns is KPI.}

\item{type}{Data will preprocess as discrete or continuous}

\item{index_column}{Column name of company id}

\item{time_column}{Column name of time value}

\item{time_values}{Time values by year}

\item{continuous_kpi_names}{Column names of continuous dynamic KPI}

\item{continuous_static_kpi_names}{Column names of continuous static (un-change over time) KPI}

\item{discrete_static_kpi_names}{Columns names of discrete static (un-change over time) KPI}

\item{current_layers}{Maximum number of layers in bayesian network, equal number of time values}

\item{desire_layers}{Number of layers for training bayesian network, less than or equal current layers}

\item{quantile_number}{For type is "discrete", number of quantile for each continuous variable}

\item{na_omit}{Is omit NA or not}

\item{debug}{Debug mode}
}
\value{
A processed data frame with pre-process params
}
\description{
Pre-processing data before training with DBN
}
\examples{
data("data_big")
type <- "continuous"
index_column <- "MST"
time_column <- "NAM"
time_values <- 2017:2019
continuous_kpi_names <- paste("K_C_V0", 1:9, sep = "")
continuous_static_kpi_names <- c("TUOI", "VON")
discrete_static_kpi_names <- c("NGANHKT")
current_layers = 3
desire_layers = 2
quantile_number = -1
head(data)
preprocessed <- preprocess_training_data(data_big, type, index_column, time_column, time_values,
                                         continuous_kpi_names, continuous_static_kpi_names,
                                         discrete_static_kpi_names, current_layers, desire_layers, quantile_number)
}
